+++ configs/module.config.php
@@ -1,6 +1,6 @@
 <?php
 
-use Zend\Mvc\Application;
+use Laminas\Mvc\Application;
 
 return [
     'service_manager' => [
@@ -13,8 +13,8 @@ return [
             'AsseticBundle\Service'       => 'AsseticBundle\ServiceFactory',
             'Assetic\AssetWriter'         => 'AsseticBundle\WriterFactory',
             'AsseticBundle\FilterManager' => 'AsseticBundle\FilterManagerFactory',
-            'Assetic\AssetManager'        => 'Zend\ServiceManager\Factory\InvokableFactory',
-            'AsseticBundle\Listener'      => 'Zend\ServiceManager\Factory\InvokableFactory',
+            'Assetic\AssetManager'        => 'Laminas\ServiceManager\Factory\InvokableFactory',
+            'AsseticBundle\Listener'      => 'Laminas\ServiceManager\Factory\InvokableFactory',
             'AsseticBundle\Cli'           => 'AsseticBundle\Cli\ApplicationFactory',
             'AsseticBundle\Configuration' => 'AsseticBundle\Factory\ConfigurationFactory',
         ],
@@ -22,9 +22,9 @@ return [
 
     'assetic_configuration' => [
         'rendererToStrategy' => [
-            'Zend\View\Renderer\PhpRenderer'  => 'AsseticBundle\View\ViewHelperStrategy',
-            'Zend\View\Renderer\FeedRenderer' => 'AsseticBundle\View\NoneStrategy',
-            'Zend\View\Renderer\JsonRenderer' => 'AsseticBundle\View\NoneStrategy',
+            'Laminas\View\Renderer\PhpRenderer'  => 'AsseticBundle\View\ViewHelperStrategy',
+            'Laminas\View\Renderer\FeedRenderer' => 'AsseticBundle\View\NoneStrategy',
+            'Laminas\View\Renderer\JsonRenderer' => 'AsseticBundle\View\NoneStrategy',
         ],
         'acceptableErrors' => [
             Application::ERROR_CONTROLLER_NOT_FOUND,